cleanup fixtures
