+++ tests/api/test_data_processing.py
@@ -4,10 +4,9 @@ from sqlalchemy.orm.exc import NoResultFound
 
 from inbox.models import DataProcessingCache
 
-from tests.api.base import api_client
 from tests.util.base import add_fake_message, add_fake_thread, default_namespace
 
-__all__ = ["api_client", "default_namespace"]
+__all__ = ["default_namespace"]
 
 
 def test_contact_rankings(db, api_client, default_namespace):
