+++ tests/api/test_events.py
@@ -4,10 +4,9 @@ from inbox.api.ns_api import API_VERSIONS
 from inbox.models import Calendar, Event
 from inbox.sqlalchemy_ext.util import generate_public_id
 
-from tests.api.base import api_client
 from tests.util.base import add_fake_event, calendar, db
 
-__all__ = ["api_client", "calendar", "db"]
+__all__ = ["calendar", "db"]
 
 
 def test_create_event(db, api_client, calendar):
