+++ tests/api/test_invalid_account.py
@@ -5,10 +5,9 @@ import mock
 import pytest
 import requests
 
-from tests.api.base import api_client
 from tests.util.base import db
 
-__all__ = ["api_client", "db"]
+__all__ = ["db"]
 
 
 @pytest.fixture
