+++ tests/api/test_searching.py
@@ -13,7 +13,6 @@ from inbox.search.backends.gmail import GmailSearchClient
 from inbox.search.backends.imap import IMAPSearchClient
 from inbox.search.base import get_search_client
 
-from tests.api.base import api_client, imap_api_client
 from tests.util.base import (
     add_fake_folder,
     add_fake_imapuid,
