+++ tests/api/test_sending.py
@@ -16,7 +16,6 @@ from inbox.basicauth import OAuthError
 from inbox.models import Event, Message
 from inbox.sendmail.smtp.postel import _substitute_bcc
 
-from tests.api.base import api_client
 from tests.util.base import imported_event, message, thread
 
 __all__ = ["thread", "message", "api_client", "imported_event"]
