+++ tests/api/test_validation.py
@@ -6,10 +6,9 @@ import pytest
 from inbox.api.validation import noop_event_update, valid_email
 from inbox.models import Namespace
 
-from tests.api.base import api_client
 from tests.util.base import add_fake_event, calendar, db
 
-__all__ = ["api_client", "db", "calendar"]
+__all__ = ["db", "calendar"]
 
 
 # TODO(emfree): Add more comprehensive parameter-validation tests.
