+++ tests/api/test_views.py
@@ -1,7 +1,7 @@
 # flake8: noqa: F811
 import pytest
 
-from tests.api.base import api_client, new_api_client
+from tests.api.base import new_api_client
 from tests.util.base import generic_account
 
 __all__ = ["api_client", "generic_account"]
